Start participants section and organize files
